clients/requirements.txt
paho-mqtt
termcolor